requirements.txt
tensorflow
typer
torch
stable_baselines3
yfinance
streamlit
pandas
numpy
gymnasium
matplotlib
pillow
tqdm
pettingzoo
ray
dm_tree
scikit-image
lz4
ray[tune]
gputil
sphinx==6.2.1
furo
